+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 
 * [74: Set minimum PHP version to 5.4](https://github.com/acelaya/ZF2-AcMailer/issues/74)
 * [79: Change license to MIT](https://github.com/acelaya/ZF2-AcMailer/issues/79)
+* [69: Remove deprecated methods](https://github.com/acelaya/ZF2-AcMailer/issues/69)
 
 ### 4.5.0